connectors/unique_search_proxy/web/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Web Search Proxy implementation"
authors = [{ name = "ThePhilAz", email = "rami.azouz@philico.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask[async] (>=3.1.2,<4.0.0)",
    "google-cloud-aiplatform (>=1.128.0,<2.0.0)",
    "google-auth (>=2.43.0,<3.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "httpx (>=0.28.0,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "ruff (>=0.14.7,<0.15.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
